update script.sh, functionality works, curls html and serves webpage

+++ script.sh
@@ -1,10 +1,11 @@
 #!/bin/sh
-FILE="index.html"
-if [-f $FILE]
-then
-    rm -rf $FILE
-else
-    curl https://bitbucket.org/bjgiller/evt-tech-challenge/raw/master/evt-web.html >> $FILE
-fi
+FILE=index.html
+[ -f $FILE ] && rm -rf $FILE
+# if test -f $FILE
+# then
+#     echo "EXISTS"
+#     rm -rf $FILE
+# fi
 
+curl https://bitbucket.org/bjgiller/evt-tech-challenge/raw/master/evt-web.html >> $FILE
 python3 server.py
